+++ tikv-client/src/main/java/com/pingcap/tikv/TiSession.java
@@ -29,11 +29,8 @@ import java.util.Map;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.function.Function;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 public class TiSession implements AutoCloseable {
-  private final Logger logger = LoggerFactory.getLogger(TiSession.class);
   private final TiConfiguration conf;
   private final ChannelFactory channelFactory;
   private Function<CacheInvalidateEvent, Void> cacheInvalidateCallback;
@@ -180,10 +177,6 @@ public class TiSession implements AutoCloseable {
     return res;
   }
 
-  public static TiSession create(TiConfiguration conf) {
-    return new TiSession(conf);
-  }
-
   /**
    * This is used for setting call back function to invalidate cache information
    *
@@ -191,6 +184,10 @@ public class TiSession implements AutoCloseable {
    */
   public void injectCallBackFunc(Function<CacheInvalidateEvent, Void> callBackFunc) {
     this.cacheInvalidateCallback = callBackFunc;
+  }
+
+  public static void clearCache() {
+    TiSession.sessionCachedMap.clear();
   }
 
   @Override
