+++ tikv-client/src/test/java/com/pingcap/tikv/PDMockServerTest.java
@@ -40,7 +40,7 @@ public abstract class PDMockServerTest {
             GrpcUtils.makeMember(2, "http://" + addr + ":" + (pdServer.port + 1)),
             GrpcUtils.makeMember(3, "http://" + addr + ":" + (pdServer.port + 2))));
     TiConfiguration conf = TiConfiguration.createDefault(addr + ":" + pdServer.port);
-    session = TiSession.create(conf);
+    session = TiSession.getInstance(conf);
   }
 
   @After
